requirements-test.txt
    robotframework
robotframework-seleniumlibrary
pytest
pytest-django
pytest-cov
pytest-xdist
coverage
pycodestyle
autopep8
